+++ packages/views/modals/create-workspace.tsx
@@ -1,12 +1,8 @@
 "use client";
 
-import { useRef, useState } from "react";
 import { useNavigation } from "../navigation";
 import { useImmersiveMode } from "../platform";
-import { toast } from "sonner";
 import { ArrowLeft } from "lucide-react";
-import { Input } from "@multica/ui/components/ui/input";
-import { Label } from "@multica/ui/components/ui/label";
 import { Button } from "@multica/ui/components/ui/button";
 import {
   Dialog,
@@ -14,84 +10,15 @@ import {
   DialogTitle,
   DialogDescription,
 } from "@multica/ui/components/ui/dialog";
-import { Card, CardContent } from "@multica/ui/components/ui/card";
-import { useCreateWorkspace } from "@multica/core/workspace/mutations";
 import { paths } from "@multica/core/paths";
-import {
-  WORKSPACE_SLUG_CONFLICT_ERROR,
-  WORKSPACE_SLUG_FORMAT_ERROR,
-  WORKSPACE_SLUG_REGEX,
-  isWorkspaceSlugConflict,
-  nameToWorkspaceSlug,
-} from "../workspace/slug";
+import { CreateWorkspaceForm } from "../workspace/create-workspace-form";
 
 export function CreateWorkspaceModal({ onClose }: { onClose: () => void }) {
   // This modal is full-screen, so it covers the app titlebar. On macOS desktop
   // we hide the traffic lights for its lifetime so the Back button in the top-
   // left corner isn't stolen by the native controls' hit-test. No-op elsewhere.
   useImmersiveMode();
-
   const router = useNavigation();
-  const createWorkspace = useCreateWorkspace();
-  const [name, setName] = useState("");
-  const [slug, setSlug] = useState("");
-  const [slugServerError, setSlugServerError] = useState<string | null>(null);
-  const slugTouched = useRef(false);
-
-  const slugValidationError =
-    slug.length > 0 && !WORKSPACE_SLUG_REGEX.test(slug)
-      ? WORKSPACE_SLUG_FORMAT_ERROR
-      : null;
-  const slugError = slugValidationError ?? slugServerError;
-
-  const canSubmit =
-    name.trim().length > 0 && slug.trim().length > 0 && !slugError;
-
-  const handleNameChange = (value: string) => {
-    setName(value);
-    if (!slugTouched.current) {
-      setSlug(nameToWorkspaceSlug(value));
-      setSlugServerError(null);
-    }
-  };
-
-  const handleSlugChange = (value: string) => {
-    slugTouched.current = true;
-    setSlug(value);
-    setSlugServerError(null);
-  };
-
-  const handleCreate = () => {
-    if (!canSubmit) return;
-    // The modal is only reachable from an authenticated workspace context
-    // (via the global modal registry). After creating a new workspace the
-    // user should land INSIDE it at its issues page, not in /onboarding —
-    // onboarding exists only for users with zero workspaces. Navigation is the
-    // only way to switch workspaces now (URL is the source of truth), so the
-    // push below is sufficient — no imperative store writes needed.
-    createWorkspace.mutate(
-      { name: name.trim(), slug: slug.trim() },
-      {
-        onSuccess: (newWs) => {
-          onClose();
-          // Navigate INTO the new workspace. The mutation's own onSuccess
-          // (in core/workspace/mutations.ts) runs before this callback and
-          // has already seeded the workspace list cache, so the destination
-          // [workspaceSlug]/layout will resolve newWs.slug → workspace
-          // synchronously without a loading flash.
-          router.push(paths.workspace(newWs.slug).issues());
-        },
-        onError: (error) => {
-          if (isWorkspaceSlugConflict(error)) {
-            setSlugServerError(WORKSPACE_SLUG_CONFLICT_ERROR);
-            toast.error("Choose a different workspace URL");
-            return;
-          }
-          toast.error("Failed to create workspace");
-        },
-      },
-    );
-  };
 
   return (
     <Dialog
@@ -135,49 +62,17 @@ export function CreateWorkspaceModal({ onClose }: { onClose: () => void }) {
               projects and issues.
             </DialogDescription>
           </div>
-
-          <Card className="w-full">
-            <CardContent className="space-y-4 pt-6">
-              <div className="space-y-1.5">
-                <Label>Workspace Name</Label>
-                <Input
-                  autoFocus
-                  type="text"
-                  value={name}
-                  onChange={(e) => handleNameChange(e.target.value)}
-                  placeholder="My Workspace"
-                />
-              </div>
-              <div className="space-y-1.5">
-                <Label>Workspace URL</Label>
-                <div className="flex items-center gap-0 rounded-md border bg-background focus-within:ring-2 focus-within:ring-ring">
-                  <span className="pl-3 text-sm text-muted-foreground select-none">
-                    multica.ai/
-                  </span>
-                  <Input
-                    type="text"
-                    value={slug}
-                    onChange={(e) => handleSlugChange(e.target.value)}
-                    placeholder="my-workspace"
-                    className="border-0 shadow-none focus-visible:ring-0"
-                    onKeyDown={(e) => e.key === "Enter" && handleCreate()}
-                  />
-                </div>
-                {slugError && (
-                  <p className="text-xs text-destructive">{slugError}</p>
-                )}
-              </div>
-            </CardContent>
-          </Card>
-
-          <Button
-            className="w-full"
-            size="lg"
-            onClick={handleCreate}
-            disabled={createWorkspace.isPending || !canSubmit}
-          >
-            {createWorkspace.isPending ? "Creating..." : "Create workspace"}
-          </Button>
+          <CreateWorkspaceForm
+            onSuccess={(newWs) => {
+              onClose();
+              // Navigate INTO the new workspace. The mutation's own onSuccess
+              // (in core/workspace/mutations.ts) runs before this callback and
+              // has already seeded the workspace list cache, so the destination
+              // [workspaceSlug]/layout will resolve newWs.slug → workspace
+              // synchronously without a loading flash.
+              router.push(paths.workspace(newWs.slug).issues());
+            }}
+          />
         </div>
       </DialogContent>
     </Dialog>
